Minor changes to make restart not throw disposable errors
