+++ lib/omnisharp-atom/services/apply-changes.ts
@@ -1,9 +1,10 @@
 var Range = require('atom').Range;
+import {Observable} from "rx";
 
 class Changes {
-    public static applyChanges(editor: Atom.TextEditor, response: { Changes: OmniSharp.Models.LinePositionSpanTextChange[]; });
-    public static applyChanges(editor: Atom.TextEditor, response: { Buffer: string });
-    public static applyChanges(editor: Atom.TextEditor, response: any) {
+    public applyChanges(editor: Atom.TextEditor, response: { Changes: OmniSharp.Models.LinePositionSpanTextChange[]; });
+    public applyChanges(editor: Atom.TextEditor, response: { Buffer: string });
+    public applyChanges(editor: Atom.TextEditor, response: any) {
         if (response.Changes) {
             var buffer = editor.getBuffer();
             var checkpoint = buffer.createCheckpoint();
@@ -18,6 +19,25 @@ class Changes {
             editor.setText(response.Buffer)
         }
     }
+
+    public applyAllChanges(changes: OmniSharp.Models.ModifiedFileResponse[]) {
+        return Observable.from(changes)
+            .concatMap(change => atom.workspace.open(change.FileName, undefined)
+                .then(editor => {
+                    var pane: HTMLElement = <any>atom.views.getView(atom.workspace.getActivePane());
+                    var title = pane.querySelector('.title.temp');
+                    var tab = pane.querySelector('.preview-tab.active');
+                    if (title) {
+                        title.classList.remove('temp');
+                    }
+                    if (tab) {
+                        tab.classList.remove('preview-tab');
+                        (<any>tab).isPreviewTab = false;
+                    }
+                    this.applyChanges(editor, change);
+                }))
+            .subscribe();
+    }
 }
 
-export = Changes
+export var changes = new Changes;
